Expand affiliate/coupon tracking coverage.
