+++ client/jetpack-connect/authorize.js
@@ -10,7 +10,6 @@ import page from 'page';
 import { connect } from 'react-redux';
 import { flowRight, get, includes, startsWith } from 'lodash';
 import { localize } from 'i18n-calypso';
-import urlUtils from 'url';
 
 /**
  * Internal dependencies
@@ -77,7 +76,6 @@ import {
 	isSiteBlacklistedError as isSiteBlacklistedSelector,
 } from 'state/jetpack-connect/selectors';
 import getPartnerSlugFromQuery from 'state/selectors/get-partner-slug-from-query';
-import { affiliateReferral } from 'state/refer/actions';
 
 /**
  * Constants
@@ -169,30 +167,6 @@ export class JetpackAuthorize extends Component {
 			const attempts = this.props.authAttempts || 0;
 			this.retryingAuth = true;
 			return retryAuth( site, attempts + 1, nextProps.authQuery.from );
-		}
-	}
-
-	componentDidMount() {
-		this.trackAffiliate();
-	}
-
-	/**
-	 * Track affiliate code based on the aff and cid URL params.
-	 */
-	trackAffiliate() {
-		const urlPath = location.href;
-		const parsedUrl = urlUtils.parse( urlPath, true );
-		const affiliateId = parsedUrl.query.aff;
-		const campaignId = parsedUrl.query.cid;
-		const subId = parsedUrl.query.sid;
-
-		if ( affiliateId && ! isNaN( affiliateId ) ) {
-			const hostPath = `${ parsedUrl.host }${ parsedUrl.pathname }`;
-			this.props.recordTracksEvent( 'calypso_jpc_refer_visit', {
-				flow: this.props.flowName,
-				page: hostPath,
-			} );
-			this.props.trackAffiliateReferral( { affiliateId, campaignId, subId, hostPath } );
 		}
 	}
 
@@ -729,7 +703,6 @@ const connectComponent = connect(
 		authorize: authorizeAction,
 		recordTracksEvent: recordTracksEventAction,
 		retryAuth: retryAuthAction,
-		trackAffiliateReferral: affiliateReferral,
 	}
 );
